+++ api/routes/categoryRouter.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+const { categoryController } = require('../controllers');
+
+router.get('/:categoryId', categoryController.categoryInfo);
+
+module.exports = router;
