+++ api/routes/index.js
@@ -2,9 +2,9 @@ const express = require('express');
 const router = express.Router();
 
 const userRouter = require('./userRouter');
-const listRouter = require('./listRouter');
+const categoryRouter = require('./categoryRouter');
 
 router.use('/users', userRouter);
-router.use('/lists', listRouter);
+router.use('/categories', categoryRouter);
 
 module.exports = router;
