--- api/routes/listRouter.js
@@ -1,8 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-const { listController } = require('../controllers');
-
-router.get('/:categoryId', listController.listInfo);
-
-module.exports = router;
